Add Makefile
This adds the Makefile. It provides install and uninstall targets which
should put both the ec2-backup program and it's manpage in their correct
places.

The install/uninstall location can further be tweaked by specifying
either DESTDIR or PREFIX (or both, or even BINPREFIX and MANPREFIX) on
the command line as well. For example: `DESTDIR=here make install` will
install to: 'here/usr/local' instead of the root.