Add main getopts structure
This sets up the initial structure of the ec2-backup program. It adds
getopts parsing for the supported option flags including usage and help
messages. Currently all it does is set variables for whatever options
are provided. The required 'dir' parameter is also checked (separately
from getopts as it does not have a flag) and stored in a variable.

This also adds the 'echoerr' function, which wraps the builtin echo
sending it's output to STDERR instead of STDOUT.