Fix formatting and variable quoting
This removes some tabs. It also quotes variables to avoid any unintended
expansion. Also it changes out `while [ 1 ]` for `while true` as I think
it's a bit more clear and also recommended by Shellcheck (SC2162) [1].

Also, this changes the quoting of the dd portion of the ssh command so
that DEVICE will be expanded before passing it to ssh so that the
intended value gets used on the remote machine.

[1]: https://github.com/koalaman/shellcheck/wiki/SC2161