Add initial test script
Automated testing is good, this adds a test script (and Makefile target
to run it). Currently it only checks the usage and help messages.

The ec2-backup usage message was updated to use basename to print it's
name instead of just $0. This was done to make the messages easier to
check in the test script and has the additional benefit of making the
output more sane.

This also adds the test directory 'sandbox' to a gitignore file in case
anything goes wrong with the test script and it leaves behind any
garbage.