Fix checks for variable flags
The CREATED_KEYPAIR and CREATED_SECURITY_GROUP variables were being
checked directly causing sh to try executing them and checking for
return codes, which left the security group and key pair not being
cleaned up. This actually checks the value of the variables making the
cleanup function to work as intended.