Add removing created volumes in case of error
This adds a check for whether or not we created a volume as well as a
parameter to the cleanup function.

Now whenever calling cleanup, if we are cleaning up from an error an
argument should be provided to cleanup. This argument is used along with
CREATED_VOLUME to determine whether or not we should delete a volume. I
have used 'cleanup error' throughout the code, although the value of the
argument doesn't matter at this point.

Note: The final use of cleanup (successful exit) does not include an
argument.